requirements.txt
Brlapi==0.6.1
Cython==0.25.2
Mako==0.9.1
MarkupSafe==0.18
Pillow==2.3.0
PyYAML==3.12
Unidecode==0.04.19
appdirs==1.4.0
apturl==0.5.2ubuntu4
asyncio==3.4.3
autobahn==0.14.1
awscli==1.10.63
botocore==1.4.53
chardet==2.2.1
checkbox==0.17.7.dev
checkbox-ng==0.3
checkbox-support==0.2
colorama==0.2.5
command-not-found==0.3
defer==1.0.6
devscripts==2.14.1ubuntu0.1
diff-match-patch==20121119
dirspec==13.10
docutils==0.12
ecdsa==0.13
ed25519==1.4
feedparser==5.1.3
friends==0.1
funcy==1.7.2
graphenelib==0.4.8
html5lib==0.999
httplib2==0.8
jmespath==0.9.0
kazam==1.4.5
language-selector==0.1
louis==2.5.3
lxml==3.3.3
matplotlib==1.3.1
mnemonic==0.15
nose==1.3.1
numpy==1.8.2
oauthlib==0.6.1
onboard==1.0.1
oneconf==0.3.7.14.04.1
pbkdf2==1.3
piston-mini-client==0.7.5
plainbox==0.5.3
prettytable==0.7.2
protobuf==3.1.0.post1
pyasn1==0.1.9
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
pylibscrypt==1.5.3
pyparsing==2.0.1
python-apt==0.9.3.5ubuntu2
python-bitsharesrpc==0.1
python-dateutil==2.0
python-debian==0.1.21-nmu2ubuntu2
python-frontmatter==0.2.1
pytz==2012c
pyxdg==0.25
requests==2.10.0
rsa==3.4.2
s3transfer==0.1.3
screen-resolution-extra==0.0.0
scrypt==0.7.1
semver==2.7.5
six==1.10.0
software-center-aptd-plugins==0.0.0
stakemachine==0.0.1
steem==0.4.2
steem-piston==0.2.3a
thin-client-config-agent==0.8
tornado==3.1.1
txaio==2.5.1
ubuntu-drivers-common==0.0.0
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
unity-scope-audacious==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-clementine==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-gmusicbrowser==0.1
unity-scope-gourmet==0.1
unity-scope-guayadeque==0.1
unity-scope-manpages==0.1
unity-scope-musique==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.7.1
usb-creator==0.2.23
virtualenv==15.0.2
websocket-client==0.37.0
websockets==2.0
wheel==0.24.0
xdiagnose==3.6.3build2
xkit==0.0.0
